requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
Pillow==10.4.0
streamlit==1.39.0
